Modified Config class to load multiple settings files
Modified the Config class to support loading multiple settings files instead of a single file. The class now checks for the existence of multiple files in the directory hierarchy and loads all the existing files. The GetValue method has been modified to search for values in all the loaded files.